Move lambda_response.py to frontend.
